Fix scheduler error retry logic (#20380)
Summary:
When we added logic to track the last iteration time in memory across
multiple ticks, we set it to the wrong value on the failure case when
the next tick shoudl retry. It's set to the current time, not the
expected schedule time - the former is typically slightly after the
latter, so the next tick won't consider the correct timerange. This PR
fixes that issue and adds a test case.

Test Plan: Adjusted test case that was now failing before the change
here

## Summary & Motivation

## How I Tested These Changes