use AssetSelection for parsing selections in asset CLI (#20383)
## Summary & Motivation

Our `dagster asset` CLI `materialize` and `list` commands both support a
`--select` option, which accepts a string. This PR makes the underlying
implementation of that build on top of `AssetSelection`. This means
that, when we improve string-based asset selection, the CLI will be able
to take advantage here.

The motivation here is supporting tags in string-based asset selection.

## How I Tested These Changes