[external-assets] Make backfill retry handle InvalidSubsetError (#20427)
## Summary & Motivation

This changes addresses a very subtle issue with functionality addressed
in https://github.com/dagster-io/dagster/pull/19091. That PR addresses
the situation where, in the middle of a backfill, the targeted job is
changed to no longer include the targeted asset. It solves this problem
by examining the execution plan returned by `get_external_job` to see if
it actually targets the requested asset-- if it does not, then the asset
graph refreshes.

This approach was failing against changes to job subsetting introduced
upstack because trying to subset a job with an asset that is not
targeted by that job now throws an InvalidSubsetError instead of
returning a "null" execution plan. This was also true before, but would
never happen for base asset jobs because all other assets were included
as source assets-- and the only reason we checked source assets is
because observable sources could be in there. Upstack, we now only check
the assets "targeted" by a job when subsetting (observables or
materializables) and ignore assets used for loading only. So we now get
a DagsterInvalidSubsetError. I also think the execution plan approach
also would've failed if targeting a non-base-asset job even prior to
this PR, it might just have been a case we weren't testing for.

In any case, the change here is just to catch a
DagsterInvalidSubsetError thrown by the code location and treat it in
exactly the same way as when the execution plan doesn't match according
to the existing logic.

## How I Tested These Changes

Existing test suite. This can be considered a refactor. It is tested by
upstack changes that throw the `DagsterInvalidSubsetError`.