[external-assets] Remove SourceAsset from AssetLayer (#20415)
## Summary & Motivation

Remove `SourceAsset` from `AssetLayer`, replacing any source assets
passed in to `AssetLayer` construction with externals. This also entails
a few `AssetLayer` callsite changes, as the existence of an assets def
for a key in the `AssetLayer` was being used as an indication that it is
executable, which is no longer true after this change.

This should be seen as a stepping stone to passing an `AssetGraph` in to
the `AssetLayer`.

## How I Tested These Changes

Existing test suite.