[external-assets] Ensure assets defs are present for all asset checks in AssetGraph (#20435)
## Summary & Motivation

There is a bug in the asset graph that I surfaced in an upstack PR,
which is that the `AssetsDefinition` for an asset check isn't available
from the asset graph if an `AssetsDefinition` has been subsetted to only
a check with no assets. This refactors `AssetGraph` to support this
niche case, which opens the door to using `AssetGraph` as the basis for
`AssetLayer`.

## How I Tested These Changes

Existing test suite. The "new" functionality is tested upstack by #20405