[external-assets] Add AssetGraph subsetting, make build_asset_job use AssetGraph (#20405)
## Summary & Motivation

Cluster of related changes that allow for using `AssetGraph` as our
general representation of "collection of assets", which allows it to be
used as the source of truth for an asset job.

- Add `AssetGraph.get_subset`. This returns a new `AssetGraph`. You pass
a set of executable asset keys and asset check keys to get a subset--
the result will automatically include any parent assets as unexecutable.
- Add a `create_unexecutable_external_assets_from_assets_def` function.
This is fulfilling the same role as `AssetsDefinition.to_source_assets`.
It would be cleaner to create a single unexecutable `AssetsDefinition`
from the passed-in one, but for now we are returning multiple
`AssetsDefinition` (once for each key) since this can be implemented in
terms of the existing `.to_source_assets`.
- Change `build_assets_job` to accept an `AssetGraph` instead of
separate lists of executable and loadable assets definitions. Modify the
two callsites to pass an `AssetGraph`.

 The next step is composing this with `AssetLayer`.

## How I Tested These Changes

Existing test suite.