[ui] Allow disabling Settings left pane items (#20449)
## Summary & Motivation

Add props to allow disabling side nav items in Settings. This is similar to the behavior of our tab components.

Also fix the font-size on button items, it's currently a bit too small.

## How I Tested These Changes

In Cloud, set permissions checks to fail for a couple of Settings items. Verify that they appear as disabled, with tooltips.