TableMetadataEntries (#20424)
## Summary & Motivation

A la part of the proposal in [this
PR](https://github.com/dagster-io/dagster/pull/20077), adds a class that
holds a set of standard metadata entries that apply to assets that are
tables.

The main impacts of this PR:
- The "columns" metadata key has been renamed to
"dagster/column_schema".
- Went with "dagster/column_schema" instead of "dagster/columns" because
the latter could be interpreted as a list of column names, while
"column_schema" implies that types are included.
- The magic string "columns" and "dagster/column_schema" no longer exist
in Python. Instead, a typed attribute on the `TableMetadataEntries`
class is used.

While the idea would be to expose `TableMetadataEntries` publicly
eventually, this PR doesn't include any public-facing API changes. It
modifies the dbt integration to use the private API.

Previous versions of this proposal had used "dagster.table" instead of
"dagster", but there's not actually a package named "dagster.table", so
stuck with "dagster".

## How I Tested These Changes