[asset-checks] Stub out AssetChecksDefinition (#20446)
## Summary & Motivation

Internal companion PR: https://github.com/dagster-io/internal/pull/8711

Implement `@asset_check` in terms of `AssetsDefinition` and convert
`AssetChecksDefinition` to a trivial subclass of `AssetsDefinition`.
This allows the consolidation of many internal codepaths and is closer
to the way we handle asset observation routines (i.e. there is no
`AssetsObserveDefinition`).

The repository/definition-level API is left in place in this PR, though
we may wish to change it in the future. `AssetChecksDefinition` are now
in a similar position to `SourceAsset`, where the symbol remains part of
our public API but is not used in the guts of our system.

## How I Tested These Changes

Existing test suite.