Skip amp evaluation on in progress asset (#20498)
Skip sending a run from AMP if there is already a run in progress for
the particular asset.

Optimized for efficiency here, so we only look at the most recent run
for an asset. Doesn't take into account the fact that there could be an
in progress run which is not the most recent, and also doesn't take into
account partitions. Those can come later, given that this is not public
for now. Just to be safe, I'm also erroring in the case of partitioned
assets.

How I tested this
Added a scenario for the case where there's an in progress asset, and
the AMP daemon would otherwise kick off a run, but doesn't as a result
of this rule being set.
Also demonstrates the existing behavior when this rule is not set, where
we automatically kick off additional runs, even if there are in progress
runs.