[ui] Repair "Show JSON" dialog stacking (#20539)
## Summary & Motivation

Reimplement the log expansion dialog to use `Dialog` directly, instead
of `showCustomAlert`. This allows us to show the "Show JSON" dialog on
top of it correctly when opened from the expansing dialog.


https://github.com/dagster-io/dagster/assets/2823852/6832c44f-fb0c-4092-a34e-0c7c4c13b0f8

## How I Tested These Changes

Find a log that needs to be expanded. Click on it, then click "Show
JSON". Verify correct dialog stacking.