Bypass slow apollo cache normalization (#20524)
## Summary & Motivation

The default cache policy for apollo queries creates caches for queries
that implicitly subscribe to updates to the underlying data. The problem
is these subscriptions are very naive (they don't seem to check if the
objects actually changed (that would require JSON.stringifying them and
comparing them)), making the process potentially extremely slow for
queries that return arrays of unbounded size.

In the case of the asset catalog we have the main query which fetches
the array of assets, and then we have granular queries done by the
virtualized asset rows themselves. These granular queries happen as the
assets load in and as each one loads in, the data gets normalized with
the the cache data for the root asset query and it's extremely slow each
time.

To circumvent the above subscription in the cache we use `fetch-policy:
'no-cache'` and we cache the data ourselves.

## How I Tested These Changes
shadow-dagit into a customer with 200k+ assets and see that their asset
catalog is usable.