Freshness checks factories (#20377)
Adds two factory methods to produce freshness checks from a list of
passed-in assets. One factory for time window partitioned assets, and
one for non partitioned assets.
More than likely, in non-time-window cases, users will want a check per
partition, which is currently not possible but will soon be possible
with the multi_asset_check decorator.
Originally, this PR had both of these within one factory, which proved
to be very hard to parse.

I opted to go for only a cron_schedule parameter for the partitioned
version, and make it required. The cron schedule is the most flexible
parameterization, since it allows for completely independent control of
the freshness condition w.r.t. the underlying partitions, but also
allows you to account for things like daylight savings. The reason it's
required is that without it, the underlying partition will always be
late immediately after the time window ends, which is confusing
behavior.

The one thing I'm not too happy about is that we can only catch certain
errors with the passed in errors (ie, partitioned asset passed into non
partitioned method) at runtime. But I don't think this needs to block
implementation.

How I tested this
Added unit tests for the following cases
- All possible parameterizations / making sure definition-time error
paths are exercised
- Combinations of cron and lag params for non partitioned assets
- Cron + partitioned assets
- Both observation and materialization events
- Incorrect assets passed to builders, and the runtime errors that
ensue.

---------

Co-authored-by: Sandy Ryza <sandy@elementl.com>