chore(pyright): rebuild pyright pins on each buildkite build (#20440)
## Summary & Motivation
Prevent errors like https://github.com/dagster-io/dagster/pull/20406
from happening by always attempting to regenerate the pyright pins on
each build.

I could put this in a separate step and scope it down to only run if any
`requirements.txt` has changed in the `pyright/` subdirectory, but want
to see if the build time difference is negligible because of `uv` first.

## How I Tested These Changes
bk