make target materialization None for check evaluations with asset observations (#20581)
## Summary & Motivation

Prior to this PR, if an asset check executed on an observable source
asset, the check evaluation's "Target materialization" would point to
the latest observation event for the asset. This was
confusing/misleading in the UI. This change causes it to be `None`
instead.

It could make sense for asset check evaluations to track observations
they correspond to, but this PR leaves that for future work, because I
think there's some difficult nuance there.

## How I Tested These Changes

Added a test. Verified that "Target materialization" doesn't show up in
the check evaluation UI for this example, and that it did prior to this
PR:

```python
from dagster import (
    AssetCheckResult,
    Definitions,
    ObserveResult,
    asset_check,
    observable_source_asset,
)


@observable_source_asset
def asset1():
    return ObserveResult()


@asset_check(asset=asset1)
def check1():
    return AssetCheckResult(passed=True)


defs = Definitions(assets=[asset1], asset_checks=[check1])
```