[ui] Search styles (#20623)
## Summary & Motivation

Make a handful of changes to asset search and global search. Worked
through these in a zoom with @braunjj. Verified all of the changes in
light and dark mode.

- Separate some asset search components from global search components.
This should hopefully allow faster iteration and experimentation on
asset search without also requiring lots of global search testing and
changes.
- Remove filter header code from global search, since it's not used
there.
- Update box-shadow styles on inputs.
- Modify border-radius values to keep input, results, and containers
flush.
- Connect asset search input to its results, apply a background to the
empty state.
- Add a "clear" X button to asset search when the input has a value.

## How I Tested These Changes

In light and dark mode, interact with global search and asset search.

- Verify default and hover states on empty and populated inputs.
- Verify correct rendering of results and empty states.
- Verify that clearing the asset search input behaves correctly.