ScheduleDefinition: Unhelpful `DeprecatedWarning` (#20523)
## Summary & Motivation
There is a `DepreciationWarning` for the `ScheduleDefinition`
constructor that triggers regardless of the inputs given.

Given the `DepreciationWarning` indicates the depreciation of the use of
`environment_vars` in the constructor, expected use would no emitting of
the warning when the `environment_vars` is unused.

**Cause**: `check.opt_mapping_param` returns `dict()` when the object
passed into it is `None`. This is counter to to the `@deprecated`
decorator which required a value of `None` to not emit the warning.

## How I Tested These Changes

This showed up in our local usage of dagster in an attempt to find
`DeprecationWarning`s across the codebase, in which this one we are
unable to resolve due to the inconsistent behavior mentioned above.

This PR includes a
`test__empty_schedule_definition__no_raise_DepreciationWarning` which
fails upon any `DeprecationWarning` raised when constructing an empty
`ScheduleDefinition`.