RFC: Add asset key set to step skipped and failed events
Summary:
The idea is that this would allow us to react, either synchronously or asynchronously, to inbound step failure and skip events, and update our understanding of the status of the asset keys that were intended to be materialized by this step. If a materialization event already came in, great. If not, mark the relevant assets as skipped or failed in the relevant table for that asset.