feat(dbt): do not execute dbt in `--debug` by default (#20660)
## Summary & Motivation
Reverts https://github.com/dagster-io/dagster/pull/19646 and
https://github.com/dagster-io/dagster/pull/19498.

Although we were getting more metadata from dbt, this metadata was:
- Prone to inaccuracy (e.g. see
https://github.com/dbt-labs/dbt-bigquery/issues/602)
- Caused pipelines to potentially double in duration, due to the
abundancy of logs

This metadata is easy to get if we introduce a wrapper (and then
metadata retrieval will be square in the ownership of this integration).
With that in mind, we don't need to rely on the metadata from `--debug`.
So remove it.

## How I Tested These Changes
pytest