[external-assets] Make stub assets part of default group (#20563)
## Summary & Motivation

Assets specified via `deps` currently have a default `group_name` of
`None`. This leads to a situation where they display as "outside of the
group" (`foo_source` in below image) when viewing the default asset
group, whereas when they are specified as `SourceAsset` or explicitly as
external assets they get a default group name of `DEFAULT_GROUP_NAME`
(and are shown like `bar_source` in the below image).


![image](https://github.com/dagster-io/dagster/assets/1531373/cee51481-31c6-4b07-99eb-b63591da17b6)

This PR makes it so that all assets, no matter how they are specified,
get a default group name of `DEFAULT_GROUP_NAME`. It also changes the
GQL schema to always return a string (instead of nullable string for
`groupName`).

cc @bengotow re this comment:
https://github.com/dagster-io/dagster/pull/20474#pullrequestreview-1936853177

## How I Tested These Changes

Existing test suite.