feat: use `difflib` by default to display `AssetKey` mismatch errors (#20664)
## Summary & Motivation
I was pursuing `sqlglot` code and saw this:
https://github.com/tobymao/sqlglot/blob/a18444dbd7ccfc05b189dcb2005c85a1048cc8de/sqlglot/dialects/dialect.py#L343-L348

And realized there was a string similarity library in the Python stdlib:
[`difflib`](https://docs.python.org/3/library/difflib.html). It's
inefficient, but does the job. Let's use that if `rapidfuzz` is missing.

Also, once https://github.com/rapidfuzz/rapidfuzz-rs introduces Python
bindings, we could probably use that instead and not have to worry about
windows dependency shenanigans.

## How I Tested These Changes
pytest: removed skips on existing tests if `rapidfuzz` is uninstalled.