[embedded-elt][sling] passing translator and replication config from decorator using metadata (#20564)
## Summary & Motivation

The current implementation of the Sling integration requires passing the
_DagsterSlingTranslator_ and replication config to both the
`@sling_assets` decorator, and the `replicate` method in the function
body.

Following the approach of `dagster-dbt` and the soon-to-be _dlt_
integration. this pull request demonstrates how we can pass the
translator to the `replicate` method through the use of metadata. This
workaround leads to a more intuitive end-user experience, albeit with
more complex implementation.

The negative to this approach is that the _translator_ and
_replication-config_ remain on the metadata, see the modified unit test.

## How I Tested These Changes

Ran unit tests.