fix(dbt): retrieve schema from sources to compute column lineage (#20682)
## Summary & Motivation
Retrieving the schema for dbt sources to input into `sqlglot` for column
lineage was untested. As a result, a syntax error in the
`log_column_level_metadata` macro went through.

Fix that, and add an explicit test where we compute lineage properly,
when a dbt source is present and is an upstream reference to a model.

## How I Tested These Changes
pytest