feat(sidebar): remove metadata button (#20702)
## Summary & Motivation
This button is rendered so oddly on the sidebar (under the description
tab), and actually corresponds to the op definition metadata.

This should be removed and merged with the existing metadata section of
the sidebar, which matches the new asset details page.

Currently, this PR does the former. If we're in agreement, I can change
this to also accommodate the latter and merge that metadata into the
existing section.

## How I Tested These Changes
local