DagsterModel.model_copy (#20736)
## Summary & Motivation

It's often useful to create a copy of an object with a few fields
replaced. We do this fairly often with `NamedTuple._replace`. When we
switch to `DagsterModel`, what should we use instead?

Pydantic 1 has `copy`. `copy` is deprecated in Pydantic 2. Pydantic 2
has `model_copy`. We support both versions.

Options:
- Define `_replace` on `DagsterModel`. This would make refactoring from
`NamedTuple` to `DagsterModel` the easiest. But it's the least
pydantic-y.
- Standardize on `copy`. This would require some way of suppressing
deprecation warnings for Pydantic 2.
- Standardize on `model_copy`. This requires overriding `model_copy` to
delegate to `copy` when using Pydantic 1.

The third option is what's implemented this PR (cribbed from
@maximearmstrong's earlier implementation of `AssetSelection.replace`).
I like it because it's the most forward-looking.

## How I Tested These Changes