feat(dbt): support singular tests with single dependency (#20698)
## Summary & Motivation
Resolves https://github.com/dagster-io/dagster/issues/18595.

1. Singular tests with a single dependency are now automatically
captured as asset checks.
2. Singular tests with multiple dependencies are only captured as asset
checks if there is an asset key configured on them. That asset key will
be used in the asset check, while the other dependencies are treated as
additional deps.

A more ideal implementation of (2) would be to specify the "attached
node" for the singular test using dbt notation, as the asset key may not
be known ahead of time.

E.g. currently, this PR will want this to be populated for (2):

```yaml
meta:
  dagster:
    asset_key: ["my_asset_key"]
```

But instead, we should allow

```yaml
# Specify dbt source as the attached node
meta:
  dagster:
    source:
      source_name: ...
      table_name: ...

# Specify dbt ref as the attached node 
meta:
  dagster:
    ref:
      ref_name: ...
      version: ...
      package_name: ...
```

But that's a bit more involved to do, so go with this approach for now.

## How I Tested These Changes
pytest