replace "__dagster_no_value" with the empty string (#20707)
## Summary & Motivation

I'm not 100% convinced this is the right direction, but this PR is meant
move forward the discussion from this slack thread:
https://dagsterlabs.slack.com/archives/C04PD3G2F7H/p1711139737671339.

It makes the case for replacing "__dagster_no_value" with the empty
string.

Why the empty string vs. `None` or `True`? A couple main reasons:
- Avoids backcompat issues from expanding the set of types tags can have
- More straightforward to build UI for inputting
- Docs come out simplest


### Avoids backcompat issues from expanding the set of types tags can
have

- Ideally, we'd like to have a single way of handling tags across the
system, i.e. run tags and op tags should work the same as asset tags.
- Run tag and op tag values are currently strings (dicts can be supplied
to tags arguments, but they're coerced to strings).
- Supporting non-string run & op tag values would be a breaking change,
because public methods like OpDefinition.tags would be able to return a
different type.

### More straightforward to build UI for inputting

We currently have the ability to run tags in the UI, and we'll probably
want to expand this in the future to asset tags at some point.

<img width="732" alt="image"
src="https://github.com/dagster-io/dagster/assets/654855/559f30c9-fad5-43c1-ae3b-4b594de4e6d0">

If `True` or `None` have special meaning, then we need to worry about
users' expectations when they input those strings directly. I suspect
there will be a non-negligible set of users who input tags both in the
UI and in code.

### Docs come out the simplest

'Tags are key/value pairs, where the keys are strings and the values are
either strings or boolean `True`/`False`. When a tag’s value is True,
only the key will be shown in the UI. E.g. `{"PII": True}` will render
as just “PII”. When the value is False, the tag will be ignored.'

vs.

'Tags are key/value pairs of strings. If a tag’s value is the empty
string, only the key will be shown in the UI. E.g. `{"PII": ""}` will
render as just “PII”.'