[refactor] Replace StepOutputVersionData with custom serialization (#13083)
## Summary & Motivation

`StepOutputVersionData` is a class that represents key-value pairs
of`(StepOutputHandle, version)`. It's part of the old memoization
system. The only reason it exists is to make a `Dict[StepOutputHandle,
str]` serdes-friendly, since non-string keys fail the default
serialization logic. Whenever we want to access the data encoded by a
list of `StepOutputVersionData`, we convert it to a dictionary form.

This PR deletes `StepOutputVersionData` in favor of a custom field
serializer (which option did not exist when `StepOutputVersionData` was
written. This change allows a `Dict[StepOutputHandle, str]` to be
serialized. While overall code volume is not reduced, it pushes
serialization details to the serdes layer where they belong.

## How I Tested These Changes

New unit test.