[ui] Fix runs launched from op selection runs (#20772)
## Summary & Motivation

When re-executing a run with an op selection, the Launchpad incorrectly
discards the op selection. This appears to be because we aren't setting
a `solidSelectionQuery` value for the session, which leads to the op
selection state being `*`.

Repair this by setting `solidSelectionQuery` based on the
`solidSelection` value returned by `ConfigForRunQuery`.

I also moved the session creation step into `onCompleted`, because it
appears that we can inadvertently kick off an infinite effect loop by
saving the session. The storage data returned by the save can re-run the
effect, the save occurs within the effect, and so on. Instead, we will
now just save the session when the query completes with a `Run` object.

## How I Tested These Changes

Run `hammer` job with only `chase_giver` selected. View the create run,
then open that run in Launchpad. Verify that the op selector is
correctly populated to `chase_giver`, and execute another run from
there. Verify that this run also has the correct op selection.