feat(dbt): support singular tests with `ref` meta (#20801)
## Summary & Motivation
A follow up to https://github.com/dagster-io/dagster/pull/20698.

Support the following syntax for configuring a dbt `ref` as the attached node:

```yaml
meta:
  dagster:
    ref:
      name: ...
      version: ...
      package_name: ...
```

This way, a user does not need to know what an a dbt resource's `AssetKey` is ahead of time when
they are configuring their singular tests as Dagster asset checks.

## How I Tested These Changes
pytest