fix(dagster-polars): ignore decimal in delta/parquet tests (#20897)
## Summary & Motivation
Decimals in Polars are officially still experimental so should not be
included in CI tests. Decimal in Delta-RS at the moment is slightly
broken, I have merged some PRs couple days ago to fix most issues but
still waiting on an upstream fix in arrow-rs, in the mean time we should
ignore roundtripping decimals there as well.