Replace RawMetadataMapping with Mapping[str, Any] in AssetOut and AssetSpec
## Summary & Motivation

Right now the `metadata` argument on `asset` is different from `AssetOut` and `AssetSpec`, which is inconsistent. You get type errors when passing vanilla Python objects as metadata dictionary values to `AssetOut` and do not when passing to `asset`. We should be consistent.

Indeed I was confused how this works with dbt as it passes in vanilla objects, and when I went to look I found a `type: ignore`. We should just loosen the type constraint to acknowledge the reality of the interface

## How I Tested These Changes

BK