Lazily evaluate build_asset_job in the run worker for the specific job being executed
Summary:
We have some profiling data that suggests that build_asset_job can be the bottleneck when loading code with many different partitions definitions. Similar to https://github.com/dagster-io/dagster/pull/19313, which moved the expensive part of evaluation of certain jobs to a lambda, move the building of each of the base asset jobs to a lambda as well.

Test Plan: BK, verify with a log call in build_asset_job that when launching a run, only build_asset_job on its particular run is evaluted.